Graduate ConvNeXt to main TorchVision area.
