+++ torchvision/models/__init__.py
@@ -1,4 +1,5 @@
 from .alexnet import *
+from .convnext import *
 from .resnet import *
 from .vgg import *
 from .squeezenet import *
